tail-cli/requirements.txt
PyYAML==5.4.1
requests==2.26.0
kafka-python==2.0.2